refactor(upload-tool): 重构i18n、命名与配置
