+++ src/gadgets/upload-tool/(meta).ts
@@ -3,7 +3,7 @@ import type { GadgetMeta } from '@/tools/gadget'
 export default {
 	withResourceLoader: true,
 	defaultEnabled: false,
-	dependencies: ['mediawiki.api', 'mediawiki.util'],
-	availableFor: { rights: ['upload'] },
+	dependencies: ['ext.gadget.site-lib', 'mediawiki.util'],
+	availableFor: { rights: ['upload'], namespaces: [-1] },
 	type: 'general',
 } satisfies GadgetMeta
