+++ src/gadgets/upload-tool/api.ts
@@ -1,7 +1,6 @@
 export const createApi = (): mw.Api =>
 	new mw.Api({
 		parameters: {
-			uselang: mw.config.get('wgUserLanguage'),
 			errorformat: 'plaintext',
 		},
 	})
